requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.62
botocore==1.35.62
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
jmespath==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
